requirements.txt
# Core Dependencies
pytest==7.4.3
python-dotenv==1.0.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.4
tqdm==4.66.1
pytz==2023.3
faker==20.1.0
ratelimit==2.2.1

# Financial Data Sources
sec_api==1.0.17
finnhub-python==2.4.18
akshare==1.12.0
tushare==1.2.89
yfinance==0.2.28

# Document Processing
marker-pdf==0.2.17
pdf-kit==1.0.0
unstructured==0.5.6
parsel==1.8.1

# Social Media & Web Scraping
praw==7.7.1

# AI/ML Dependencies
langchain==0.1.0
openai==1.6.1

# FinNLP - Install from GitHub
git+https://github.com/AI4Finance-Foundation/FinNLP.git
